pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = 'setuptools.build_meta'

[tool.setuptools.packages.find]
where = ["."] 
include = ["pyre"]

[project]
name = "pyre"
version = "1.5.2" 
authors = [
    { name = "James Anderson", email = "James.R.Anderson@utah.edu" },
]
description = "A helper library that wraps python threads, multiprocessing, a process pool for shell commands, and parallel python with the same interface"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
dependencies = [
    "dependency-injector >= 4.45.0",
    "six >= 1.16",
    "numpy >= 1.26",
    "matplotlib >= 3.8",
    "rtree>=1.3",
    "PyOpenGL>=3.0",
    "pillow>=2.3",
    "pydantic >= 2.9.2",
    "PyYAML>=6.0.2",
    "wxPython>=4.0",
    "nornir_shared @ git+https://github.com/jamesra/nornir-shared.git@dev-v1.5.2",
    "nornir_pools @ git+https://github.com/jamesra/nornir-pools.git@dev-v1.5.2",
    "nornir_imageregistration @ git+https://github.com/jamesra/nornir-imageregistration.git@cupy-v1.6.5",
    "nornir_buildmanager @ git+https://github.com/jamesra/nornir-buildmanager.git@dev-v1.6.5",
]
classifiers = [
    "Programming Language :: Python :: 3.13",
 
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
]

[project.optional-dependencies]
test = ["nose"]
gl = ["pyopengl-accelerate"]


[project.urls]
Homepage = "https://github.com/jamesra/nornir-pyre"
Issues = "https://github.com/jamesra/nornir-pyre/issues"

[project.scripts]
pyre = "pyre.__main__:main"